add error handling

+++ src/main/java/com/helmigandi/app/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.helmigandi.app.student;
 
+import com.helmigandi.app.student.exception.BadRequestException;
+import com.helmigandi.app.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,11 +19,19 @@ public class StudentService {
 
     public void addStudent(Student student) {
         // check if email is taken
+        Boolean existEmail = studentRepository.selectExistsByEmail(student.getEmail());
+        if (existEmail) {
+            throw new BadRequestException("Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
         // check if student exists
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exits");
+        }
         studentRepository.deleteById(studentId);
     }
 }
